backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-tavily>=0.2.17",
    "langgraph-cli[inmem]>=0.4.12",
    "langserve>=0.3.3",
    "python-dotenv>=1.2.1",
    "sse-starlette==2.1.0",
]
